+++ cats/io/input/function.py
@@ -1,5 +1,3 @@
-from datetime import datetime
-from pathlib import Path
 from ray.data import Dataset
 from cats.utils import wait_for_directory
