+++ cats/utils/__init__.py
@@ -1,7 +1,4 @@
-import boto3
-import os, time, json, subprocess
-from copy import deepcopy
-from pprint import pprint
+import os, time, subprocess
 
 
 class Dict2Class(object):
@@ -18,133 +15,6 @@ def subproc_run(cmd):
         shell=True,
         text=True
     )
-
-
-def catSubmit(order_request, meshClient):
-    print("Order:")
-    order = json.loads(meshClient.cat(order_request["order_cid"]))
-    print()
-    pprint(order)
-    print()
-
-    ppost = lambda args, endpoint: \
-        f'curl -X POST -H "Content-Type: application/json" -d \\\n\'{json.dumps(**args)}\' {endpoint}'
-    post = lambda args, endpoint: \
-        'curl -X POST -H "Content-Type: application/json" -d \'' + json.dumps(**args) + f'\' {endpoint}'
-
-    post_cmd = post({'obj': order_request}, order["endpoint"])
-    print(ppost({'obj': order_request}, order["endpoint"]))
-    print()
-    response_str = subprocess.check_output(post_cmd, shell=True)
-    output_bom = json.loads(response_str)
-
-    output_bom['POST'] = post_cmd
-    return output_bom
-
-
-def flatten_bom(bom_response, meshClient):
-    invoice = json.loads(
-        meshClient.cat(bom_response["bom"]["invoice_cid"])
-    )
-    invoice['order'] = json.loads(
-        meshClient.cat(invoice['order_cid']),
-    )
-    invoice['order']['flat'] = {
-        'function': json.loads(meshClient.cat(invoice['order']["function_cid"])),
-        'invoice': json.loads(meshClient.cat(invoice['order']["invoice_cid"]))
-    }
-    bom_response["flat_bom"] = {
-        'invoice': invoice,
-        'log': json.loads(
-            meshClient.cat(bom_response["bom"]["log_cid"])
-        )
-    }
-    return bom_response
-
-
-def create_order_request(
-    ingress_subproc,
-    integrated_subproc,
-    egress_subproc,
-    integration_cache_subproc,
-    data_dirpath,
-    structure_filepath,
-    meshClient,
-    JOB_HOME,
-    endpoint='http://127.0.0.1:5000/cat/node/execute'
-):
-    structure_cid, structure_name = meshClient.cidFile(structure_filepath)
-    function = {
-        'ingress_subproc_cid': meshClient.ipfsClient.add_pyobj(ingress_subproc),
-        'integrated_subproc_cid': meshClient.ipfsClient.add_pyobj(integrated_subproc),
-        'egress_subproc_cid': meshClient.ipfsClient.add_pyobj(egress_subproc),
-        'integration_cache_subproc_cid': meshClient.ipfsClient.add_pyobj(integration_cache_subproc),
-        'infrafunction_subproc_cid': None
-    }
-    invoice = {
-        "data_cid": meshClient.cidDir(data_dirpath)
-    }
-    order = {
-        "function_cid": meshClient.ipfsClient.add_str(json.dumps(function)),
-        "structure_cid": structure_cid,
-        "invoice_cid": meshClient.ipfsClient.add_str(json.dumps(invoice)),
-        "structure_filepath": structure_name,
-        "JOB_HOME": JOB_HOME,
-        "endpoint": endpoint
-    }
-    order_request = {
-        'order_cid': meshClient.ipfsClient.add_str(json.dumps(order))
-    }
-    return order_request
-
-
-def linkProcess(
-        cat_response,
-        meshClient,
-        ingress_subproc=None,
-        integrated_subproc=None,
-        egress_subproc=None,
-        integration_cache_subproc=None,
-        infrafunction_subproc=None
-):
-    flattened_bom = flatten_bom(cat_response, meshClient)
-    flat_bom = deepcopy(flattened_bom['flat_bom'])
-    function_cids = flat_bom['invoice']['order']['flat']['function']
-    function = {}
-    if ingress_subproc is not None:
-        function['ingress_subproc_cid'] = meshClient.ipfsClient.add_pyobj(ingress_subproc)
-    else:
-        function['ingress_subproc_cid'] = function_cids['ingress_subproc_cid']
-    if integrated_subproc is not None:
-        function['integrated_subproc_cid'] = meshClient.ipfsClient.add_pyobj(integrated_subproc)
-    else:
-        function['integrated_subproc_cid'] = function_cids['integrated_subproc_cid']
-    if egress_subproc is not None:
-        function['egress_subproc_cid'] = meshClient.ipfsClient.add_pyobj(egress_subproc)
-    else:
-        function['egress_subproc_cid'] = function_cids['egress_subproc_cid']
-    if integration_cache_subproc is not None:
-        function['integration_cache_subproc_cid'] = meshClient.ipfsClient.add_pyobj(integration_cache_subproc)
-    else:
-        function['integration_cache_subproc_cid'] = function_cids['integration_cache_subproc_cid']
-    if infrafunction_subproc is not None:
-        function['infrafunction_subproc_cid'] = meshClient.ipfsClient.add_pyobj(infrafunction_subproc)
-    else:
-        function['infrafunction_subproc_cid'] = function_cids['infrafunction_subproc_cid']
-    new_function_cid = meshClient.ipfsClient.add_str(json.dumps(function))
-
-    invoice = flat_bom['invoice']
-    input_invoice = {'data_cid': invoice['data_cid']}
-    new_invoice_cid = meshClient.ipfsClient.add_str(json.dumps(input_invoice))
-
-    order = invoice['order']
-    order['function_cid'] = new_function_cid
-    order['invoice_cid'] = new_invoice_cid
-    del order['flat']
-    order['endpoint'] = 'http://127.0.0.1:5000/cat/node/link'
-
-    order_request = {'order_cid': meshClient.ipfsClient.add_str(json.dumps(order))}
-    return order_request
 
 
 def wait_for_directory(directory_path, check_interval=1):
